huge refactoring in SVG generator. Now symbols are stored in <defs> (we are using the BLOCK_HEADER handles)
But I am still not sure about the correct procedure to select the entities that must be put inside a <g>.
I tried to get all objects from first_entity until last_entity as they appear in the file, but that does not seem to be 
correct. What is the correct way to iterate over entities?

R_2004+ provides a vector of handles pointing to everything that must go inside a <g>, but prior versions only tell us the 
1st and last entities.