[w] cleaning up handle encoding
Just cleaned up handle encoding code. It still doesn't work perfectly, but that's the way to go.
There is still the big question "who's going to write metadata like handle codes and values (among other flags) when writing a new file". For now, the application. In the future, a better API. So, I'm assuming that when encoding a file the Dwg_Handle's are correctly populated and ready to be encoded to the bitstream

*src/encode.c: cleaned up dwg_encode_handleref, write supposedly existing ref->handleref