fix the dxfname, no final *
the last char is always \0, dont replace it with *.
the len BS includes the final \0, but we keep that.
we only store the ASCIIZ anyway.

also don't replace unprintable chars with ~.
we do support encode and codepages conversion later.
with codepages isprint() is relative.
thus unprintable chars can only be replaced in temp. print helpers.