WIP more 2007 handle stream
not for entities yet. there we dont have a bitsize.
initialize hdlpos after reading the object bitsize.
before 2007 hdl_dat is just the same as dat, ditto str_dat.
str_dat is a copy then because strings are interleaved.
hdl_dat just needs one advance, because the handles are only
at the end.