add CONTRIBUTING, adjust TODO
copy CONTRIBUTING from the GCC project and adjust it.  Esp.  important for
Legal Prerequisites.  This is an official and important GNU package, with
ownership assigned to the FSF, to able to protect us from legal threats.

adjust TODO, README and README-alpha.
Mention that unreleased downloads of the library will crash
as it will try to decode unsupported DWG versions. Released
downloads will rather abort then.