json: new geojson output formatter
only LINE,LWPLINE,POLYLINE_2D,INSERT (unexploded) entities so far.

TODO: facetting of curves, ocs/ucs transforms, use the new pline points api.

TODO: NOCOMMA and \n not with stdout. stdout is line-buffered
so NOCOMMA cannot backup past the previous \n to delete the comma.
We really have to add the comma before, not after, and special case the first
field, not the last to omit the comma.