fix massive obj_string_stream bug
I once wrote it to take the bitsize as 2nd arg, and then refactored it
to take the object. (for the has_string flag).
Not after carefully readng the mingw gcc warnings, I didnt use it in the header,
but manually pasted the decl into the c files. but just the old decl!
Tahts why the string stream offset calculation was so unstable.

This should improve the r2007+ situation a lot.

Add reused functions to its headers and use it.
Add restrict to the dwg_decode_##token methods
Fix redundant redeclaration of 'dwg_decode_add_object'