api: add parent field to each _dwg_{entity,object}_ENTITY
See [GH #11].
We have to jump through absurd hurdles in the API to get around that.
The parent points to the _dwg_object_{entity, object}, which points
via the object field to the _dwg_object.

Also add parent fields to each subentity, such
as e.g. MLINE_line - > MLINE_vertex -> MLINE,
and HATCH.