2007: fix section_string_stream for overlarge data_size
The overflow logic with hi_size was wrong. Too small type.

We can now read many AEC drawings with a data_size > 0x8000
and >300 classes. E.g. Autodesk Architectural Desktop 2007.
But some of them are still problematic.

Allow reading r2010+ now with a tarball release, and enable
the xmlsuite for it. But programs/alive.text does not work yet with
Leader_2010+: dwg2dxf