dxf: add missing LINE,... pts. POLYLINE special-case
before r2004 there was no LWPOLYLINE in a DXF,
only a POLYLINE. need to convert to VERTEX_2D later.

there was special DXF DECODER/ENCODER logic in the spec
missing out on the DXF case.

simplify minimal handling on some objects on the switch,
not to run into MTEXT again