add 3BLL type as documented, enable TABLE <= r2007
BLL uses 3 bits for len, but it is documented to use 3B: 1-3 bits
stopping at the first 0 bit. Which would leave us with
0: 0, 10: 2, 110: 6, 111: 7. 100 for 4 or 101 for 5 would be invalid.
But this documentation is wrong, this encoding is not used for
REQUIREDVERSIONS, picture_size.

Add the unused documented variant as 3BLL type.

Use the old TABLE pre-2010 as STABLE, and as DEBUGGING since r2010+.
dxf: Use the entity dxfname for variable types: ACAD_TABLE.
Fixes some entities for the picture_size.
Trace more common entity data.