start with ACIS/SAT version 2
which is most likely a 5 byte link to a blob of their fork of ACIS 7.0
called now ShapeManager, with all the versions.
But all the different ACIS data has in version 2 the same content:
5: BL + 0d254c8109 + 0: BL, so there must be something else.

There is EED data attached to it, with 0-2 handles and several code 70 and one 71.
The handles are to APPID instances: ACAD_OBJECT_CHANGE_GUID and ACAD_STEPID.