configure: allow PERL, check --disable-python
allow overriding PERL=cperl5.26.1 e.g. for using ccache.
allow swig bindings for perl when python is disabled,
because we ship the generated swig_perl.c

I really need to cache cperl and it's used ccache compiler,
to be able to use ccache in the perl binding.

remove bindings/perl/Makefile.in generation. forgot that earlier.