api: add dwg_block_control, dwg_{model,paper}_space_ref
dwg_block_control is easy, always the first object,
and also dwg->block_control.
but the canonical model and paper space headers are a bit tricky to get
to iterate over it.
fix the implementations and docs, and use the new API.

the old api via dwg_obj_block_control_get_model_space from the block_control
was awful and wrong.