unknown: split up generation of static array of struct
we need 3 parts to be able to initialize it statically.
disable alldwg.inc for now. we don't need to find the fields order and pos,
only for a consistency crosscheck against all DWG's without matching DXF afterwards.

add 2 fields to the dxf fields: type and pos, resolved by the solver.