Fix -v5 overshoot errors
store the size of the common entity/object data, until the
object specific fields start. for a proper UNKNOWN_*.num_bytes size.
Note that bitsize does not include the size of the common handles, so we start
counting at obj->address, but obj->common_size starts counting after the type,
so we still overshoot by ~18