dxf: rename new ent->parent to ent->owner (#82)
its confusing both ways. the corresponding object ptr is called
parenthandle, but the obj->parent is something else: inheritance.

officially we have hard and soft owners, owning the reference,
for liveness on deletion.

in the end it would be just too confusing mixing up the _obj->parent and
obj->parent with _ent->parent. maybe rename _obj->parenthandle to ownerhandle
also.