fix #99 first SEGV
dwg_dxf_LEADER@dwg.spec:2034-3___null-pointer-dereference
This is at FIELD_3DPOINT_VECTOR in LEADER.
There are illegal size fields (num_points) in LEADER on decode, which are not yet set to 0.
Add a _LV macro variant for size lvalue, which can be set to 0.

Also found an entity with entmode 3 but empty ownerhandle.
Write an empty ownerhandle then.
Fixed also a typo in writing the mspace block_record with entmode 2: PSPACE => MSPACE