change DIMASSOC, indxf: add_DIMASSOC
change the spec a bit.
H 331 seems to be repeated with main_subent_type 73,
H 332 with intsect_subent_type 74

The other handles seem to be superfluous.
Change the order of order of some fields to allow proper dxf
roundtrips.
(just rotated_type fails still)

Only the indices with the bit set in associativity are filled. (no change)