decode: skip EED parsing on illegal bitsize
Also fail early on invalid owner handles (not 0.x.x).
TODO: We might try to reconstruct the handle (previous plus one, with known
handleoffset) and search for it, to re-align the lost position.

It's next to impossible that with an illegal size/bitsize, EED can be parsed
correctly, which is then subsequently leaked.  Try at least the other Common
EED Handles then.
Fixes fuzzed dwgrewrite leaks of EED.
See GH #191