injson: recalc eed.size on wstrings
if does_cross_unicode_datversion, i.e. from r2007+ to r2000
we need to shorten the eed size of the size frame.
encode does not do that for us, we need to do that in the importers.
in_dxf already did that by necessity.
The XRECORD.xdata_size and obj->size,bitsize are always calculated.

No more INVALIDEED check-json errors in all samples.