encode: init hdl_dat stream earlier #205
add obj_flush_hdlstream to START_HANDLE_STREAM.
initially handles are written to a fresh hdl stream.
When we reach START_HANDLE_STREAM (entity and object) we flush this to dat.
On encode we usually don't know obj->bitsize and obj->hdlpos beforehand.

This greatly simplifies the dwg.spec, avoiding 2nd loops just for
the handles, and avoiding free problems.