decode: loosen invalid handlemap offset fixup
With a lot of deleted objects the "Invalid handleoff" warning might be bogus,
max_handle (hash size) might be too small yet.
Plus the offset fixup relied on a wrong calculation of the previous size.
The CRC is included in the size.

While we are here, just warn and dont fixup the offset, in my case from a
good value to a broken one.
Fixes example_2007-2010 json roundtrip