Unify 3DSOLID layout, extend AcDbEvalExpr
ensure all 3DSOLID objects have the same layout (one is an object, not an entity).
implement the AcDbEvalExpr fields, and use it.
Not as extra struct as history_node, just inline it.
A history_node is an seldom used extension, the evalexpr is a major feature
(and already used)

AcDbEvalExpr really should have the values as union, but does not work yet.
Try later.