Update README, Fix typo in install instructions
Fixed a small typo in the README file which led to: "configure: WARNING: unrecognized options: --disable--write". Especially when copy & pasting the install instructions.
Disclaimer: This is my first contribution to a open source project and I don't know if I have done it correctly (e.g. correct branch and so on). :)