encode: create r2000 linked entity list in SEQEND
in_postprocess_SEQEND needs to link the vertex[] or attribs[] entities
when downconverting.
This is normally done in in_postprcess_handles, but prevented from
entmode=0 there.

with indxf this is not yet enabled, because dwg->header.version
is higher than R_2000 on downconvert. it is not yet set to the target version,
dat neither.