WIP add has_subclass API
for the previous indxf fix.
We need to check if a subclass belongs to a class,
or which subclasses are valid in a class.
Parse dwg.spec for all SUBCLASSes and fill in
a new dwg_type_subclasses[] array, sorted for bsearch.

We expand more defines now. Also parametric _fields ().
See GH #295