dwgfuzz: try to use sscanf_s
which glibc/bsd did not implements, because politics and they do not
care about bounds checks.

So to protect from stack overflows use %119s size bounds, and restrict
the charset.  We can do that for fuzzing, but for a real "add" language we
would need to accept utf8, and write a real parser.