indxf: reset j on non-vector fields
we have several special cases for vectors or pts, where we
need a global j counter. but when we set a single field, such as a
num_clip_verts 91 or such. then we need to reset j, otherwise
we could run into vector or pts overflows.
As in oss-fuzz issue 32755 and 32785