CodeQL #53, compare long with narrow
Comparison of narrow type with wide type in loop condition

In a loop condition, comparison of a value of a narrow type with a value of
a wide type may result in unexpected behavior if the wider value is
sufficiently large (or small).  This is because the narrower value may
overflow.  This can lead to an infinite loop.