codeql UB: pointer overflow check
When checking for integer overflow, you may often write tests like p + i <
p.  This works fine if p and i are unsigned integers, since any overflow in
the addition will cause the value to simply "wrap around." However, using
this pattern when p is a pointer is problematic because pointer overflow has
undefined behavior according to the C and C++ standards.  If the addition
overflows and has an undefined result, the comparison will likewise be
undefined; it may produce an unintended result, or may be deleted entirely
by an optimizing compiler.

fixes codeql #51, #52